Add explicit casts to silence CodeSonar coercion warnings
  parseSocketSpec already validates port range (1024–65535) before
  assignment; cast makes the narrowing to unsigned short explicit.
  IPND configService tag lengthType is a signed char by protocol
  definition; cast documents the intentional narrowing from atoi.

+++ ici/library/platform.c
@@ -2357,7 +2357,7 @@ int parseSocketSpec(char *socketSpec, unsigned short *portNbr,
 		}
 		else
 		{
-			*portNbr = i4;
+			*portNbr = (unsigned short) i4;
 			portValid = 1;
 		}
 	}
